Update db context
Se automatiza el seteo de las propiedades de baseModel.
