--- CleanArchitecture/CleanArchitecture.Data/Class1.cs
@@ -1,7 +1,0 @@
-﻿namespace CleanArchitecture.Data
-{
-    public class Class1
-    {
-
-    }
-}
