+++ CleanArchitecture/CleanArchitecture.Data/Persistence/StreamerDbContext.cs
@@ -1,0 +1,62 @@
+﻿using CleanArchitecture.Domain;
+using CleanArchitecture.Domain.Common;
+using Microsoft.EntityFrameworkCore;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace CleanArchitecture.Infrastructure.Persistence;
+
+public class StreamerDbContext : DbContext
+{
+    public StreamerDbContext(DbContextOptions<StreamerDbContext> options) : base(options)
+    {
+    }
+
+    //protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
+    //{
+    //    optionsBuilder.UseSqlServer(@"Server=(localdb)\MSSQLLocalDB;Database=Streamer;Trusted_Connection=True;")
+    //        .LogTo(Console.WriteLine, new[] { DbLoggerCategory.Database.Command.Name }, Microsoft.Extensions.Logging.LogLevel.Information)
+    //        .EnableSensitiveDataLogging();
+    //}
+
+    public override Task<int> SaveChangesAsync(CancellationToken cancellationToken = default)
+    {
+        foreach (var entry in ChangeTracker.Entries<BaseDomainModel>())
+        {
+            switch (entry.State)
+            {                
+                case EntityState.Modified:
+                    entry.Entity.LastModifiedDate = DateTime.Now;
+                    entry.Entity.LastModifiedBy = "system";
+                    break;
+                case EntityState.Added:
+                    entry.Entity.CreatedDate = DateTime.Now;
+                    entry.Entity.CreatedBy = "system";
+                    break;                
+            }            
+        }
+        return base.SaveChangesAsync(cancellationToken);
+    }
+    protected override void OnModelCreating(ModelBuilder modelBuilder)
+    {
+        modelBuilder.Entity<Streamer>()
+            .HasMany(m => m.Videos)
+            .WithOne(m => m.Streamer)
+            .HasForeignKey(m => m.StreamerId)  //Fluent API se utiliza cuadno los modelos no cuamplen las convenciones de EF para representar las relaciones.
+            .IsRequired()
+            .OnDelete(DeleteBehavior.Restrict);
+
+        modelBuilder.Entity<Video>()
+            .HasMany(m => m.Actors)
+            .WithMany(m => m.Videos)
+            .UsingEntity<VideoActor>(
+            pt => pt.HasKey(e => new { e.ActorId, e.VideoId })
+            );
+
+    }
+    public DbSet<Streamer> Streamers { get; set; }
+    public DbSet<Video> Videos { get; set; }
+}
